feat: Multi-currency infrastructure, fee precision, and calculateFee fix
- Add per-currency CONVERSION_FACTORS, CURRENCY_DECIMALS maps and helper
  methods (getConversionFactor, getCurrencyDecimals) in Constants
- Change default fee from 0.1% to 0.01% (CONTACT_DEFAULT_FEE_PERCENT)
- Change DB amount columns from INT to BIGINT for large-value currency support
- Fix calculateFee formula: was currency-dependent (only worked for USD),
  replaced with currency-independent `amount * feePercent / 100`
- Fix inconsistent fee scale between getDefaultFee() (raw %) and DB
  getFeePercent() (scaled INT) — callers now normalize before calculateFee
- Replace all hardcoded `/ 100` and `number_format(..., 2)` with per-currency
  Constants methods across GUI, CLI, and services
- Add BTC-ready skip-guarded tests for future currency activation
- Update docs and CHANGELOG

+++ files/src/database/DatabaseSchema.php
@@ -59,7 +59,7 @@ function getContactCreditTableSchema() {
     return "CREATE TABLE IF NOT EXISTS contact_credit (
         id INTEGER PRIMARY KEY AUTO_INCREMENT,
         pubkey_hash VARCHAR(64) NOT NULL,
-        available_credit INT DEFAULT 0,
+        available_credit BIGINT DEFAULT 0,
         currency VARCHAR(10),
         updated_at TIMESTAMP(6) DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
         UNIQUE INDEX idx_contact_credit_hash_currency (pubkey_hash, currency),
@@ -74,7 +74,7 @@ function getContactCurrenciesTableSchema() {
         pubkey_hash VARCHAR(64) NOT NULL,
         currency VARCHAR(10) NOT NULL,
         fee_percent INT NOT NULL,
-        credit_limit INT DEFAULT NULL,
+        credit_limit BIGINT DEFAULT NULL,
         status ENUM('accepted', 'pending', 'blocked') DEFAULT 'pending',
         direction ENUM('incoming', 'outgoing') DEFAULT 'incoming',
         created_at TIMESTAMP(6) DEFAULT CURRENT_TIMESTAMP,
@@ -89,8 +89,8 @@ function getBalancesTableSchema() {
     return "CREATE TABLE IF NOT EXISTS balances (
         id INTEGER PRIMARY KEY AUTO_INCREMENT,
         pubkey_hash TEXT NOT NULL,
-        received INT NOT NULL,
-        sent INT NOT NULL,
+        received BIGINT NOT NULL,
+        sent BIGINT NOT NULL,
         currency VARCHAR(10),
         INDEX idx_balances_pubkey_hash (pubkey_hash)
     )";
@@ -131,7 +131,7 @@ function getTransactionsTableSchema() {
         receiver_address VARCHAR(255) NOT NULL,
         receiver_public_key TEXT NOT NULL,
         receiver_public_key_hash VARCHAR(64),
-        amount INT NOT NULL,
+        amount BIGINT NOT NULL,
         currency VARCHAR(10) NOT NULL,
         timestamp DATETIME(6) DEFAULT CURRENT_TIMESTAMP,
         txid VARCHAR(255) UNIQUE NOT NULL,
@@ -248,9 +248,9 @@ function getP2pTableSchema() {
         time BIGINT NOT NULL,
         expiration BIGINT NOT NULL, /* unix epoch (micro) seconds */
         currency VARCHAR(10) NOT NULL,
-        amount INTEGER NOT NULL,
-        my_fee_amount INTEGER,
-        rp2p_amount INTEGER,  /* Total amount from RP2P response (including all relay fees) - set when awaiting_approval */
+        amount BIGINT NOT NULL,
+        my_fee_amount BIGINT,
+        rp2p_amount BIGINT,  /* Total amount from RP2P response (including all relay fees) - set when awaiting_approval */
         destination_address VARCHAR(255), /* only set if you are the original sender */
         destination_pubkey TEXT,
         destination_signature TEXT,
@@ -332,7 +332,7 @@ function getRp2pTableSchema() {
         id INTEGER PRIMARY KEY AUTO_INCREMENT,
         hash VARCHAR(255) NOT NULL UNIQUE, /*This is the hash of the final recipient address + salt + time*/
         time BIGINT NOT NULL,
-        amount INTEGER NOT NULL,
+        amount BIGINT NOT NULL,
         currency VARCHAR(10) NOT NULL,
         sender_public_key TEXT NOT NULL,
         sender_address VARCHAR(255) NOT NULL,
@@ -350,12 +350,12 @@ function getRp2pCandidatesTableSchema() {
         id INTEGER PRIMARY KEY AUTO_INCREMENT,
         hash VARCHAR(255) NOT NULL,
         time BIGINT NOT NULL,
-        amount INTEGER NOT NULL,
+        amount BIGINT NOT NULL,
         currency VARCHAR(10) NOT NULL,
         sender_public_key TEXT NOT NULL,
         sender_address VARCHAR(255) NOT NULL,
         sender_signature TEXT NOT NULL,
-        fee_amount INTEGER NOT NULL,
+        fee_amount BIGINT NOT NULL,
         created_at TIMESTAMP(6) DEFAULT CURRENT_TIMESTAMP,
         INDEX idx_rp2p_cand_hash (hash),
         INDEX idx_rp2p_cand_hash_amount (hash, amount ASC),
@@ -546,8 +546,8 @@ function getCapacityReservationsTableSchema() {
         id INTEGER PRIMARY KEY AUTO_INCREMENT,
         hash VARCHAR(128) NOT NULL,
         contact_pubkey_hash VARCHAR(64) NOT NULL,
-        base_amount INT NOT NULL,
-        total_amount INT NOT NULL,
+        base_amount BIGINT NOT NULL,
+        total_amount BIGINT NOT NULL,
         currency VARCHAR(10) NOT NULL,
         status ENUM('active', 'released', 'committed') DEFAULT 'active',
         created_at TIMESTAMP(6) DEFAULT CURRENT_TIMESTAMP(6),
